Support notes with <4 fields
The Chinese flashcards have more than 4 fields, and the `req` value
(which indicates which fields are required for a template) for the
Model class assumed that there were at least 4 fields.

Generate `req` such that any number of fields/templates is supported.
All templates are assumed to require all fields (not sure in what
situations this wouldn't be true...).