Fix multiple decks in single package
We never actually supported multiple decks in a single package. Each
deck would insert its own col row, but Anki expects col to only have a
single row (which includes the details of all the decks and all the
models).

Fix it so that we update the "models" and "decks" fields of the existing
col row in Deck.write_to_db, instead of creating a new col row.

Also add a test for this behavior.

Thanks to @sciencemanx and @marccarre for identifying the bug and
helping with this patch!