Allow setting sort_field from model
It is now possible to define a `sort_field` value for notes by
setting a `sort_field_number` when creating the model.
When the `sort_field` argument is unused upon note creation, then
the field corresponding to the models `sort_field_number` will be
used. The default value is `1` (for first field).

The changes are compatible with previous versions.