Bump to a newer version of Anki for tests
Use a more recent version of Anki for running tests. The version we now
test against is from 2019-12-24. Previously, we were testing against a
version from 2017-04-13.

We can't do anything past 2019-12-24 currently because that's the last
date when there wasn't any Rust code. I need to figure out how to build
the Rust code in CI in order to get newer versions working.