Check for invalid HTML tags in fields
If field data contains a substring matching <.*> that isn't a valid HTML
tag, Anki will still try to interpret it as an HTML tag, which will
result in messed up output. Field data should be HTML-encoded (so if
you want a literal < you should write &lt; or use html.escape() ), but
sometimes this isn't clear to users.

Add a warning that triggers when invalid HTML tags are detected in the
fields for a Note to make users aware of this problem.

Helps address https://github.com/kerrickstaley/genanki/issues/28 .