Add built-in models
Add the following genanki.Model instances:
  - genanki.BASIC_MODEL
  - genanki.BASIC_AND_REVERSED_CARD_MODEL
  - genanki.BASIC_OPTIONAL_REVERSED_CARD_MODEL
  - genanki.BASIC_TYPE_IN_THE_ANSWER_MODEL
  - genanki.CLOZE_MODEL

These can be used in the same way as Anki's built-in models and are
identical to them except for their names, which are different due to a
technical issue with how Anki handles built-in model IDs.

This also updates some fields we set on each model when we serialize a
deck to more closely match Anki's output.