Add "Back Extra" field to Cloze builtin model (#101)
Due to a mistake, the "Back Extra" field was omitted from `genanki.builtin_models.CLOZE_MODEL`. This field shows extra information on the back of the card.

Add this field to `CLOZE_MODEL`. Also, add some code that runs inside Note.write_to_db that checks if `.fields` has only one element and `.model` is `CLOZE_MODEL`, and if so, fixes `.fields` and emits a warning.

This change may cause minor backwards-incompatibility issues, because the ID of `CLOZE_MODEL` has changed. The only visible effect should be that a deck of Cloze notes created before this change and a similar deck created after this change will have different Note Types when both are imported into Anki. (Note Type is the Anki UI's term for what the source code calls Model).