Add test_furigana_field
This is related to https://github.com/kerrickstaley/genanki/pull/44 .
The test already passes, which suggests that something has changed since
the pull request was originally opened.

I thought it was related to the switch from pystache to chevron, but I
did git revert 9319f34 and re-ran test and they still passed. Shrug.