Fix test execution on Python 3.12
Python 3.12 deprecates setup.py and distutils. Update setup_tests.sh and
run_tests.sh to make them work again in 3.12.

We manually patch the Anki source code to not depend on distutils. This
fix was applied upstream but we test against an old Anki version so we
don't have to build Rust stuff. We should stop kicking that can down the
road at some point.

Note that tests currently fail both locally and in CI with the error
"AttributeError: 'TestWithCollection' object has no attribute 'col'".
I think this error is unrelated to this commit and I will fix it in a
subsequent commit.