Add support for opendoas & void-linux to setup_tests script (#145)
Some systems (like mine) do not have `sudo` installed, but use
the `opendoas` package to perform the same function.

The Void Linux package manager (xbps) was fairly easy to implement
support for, though I restructured that part of code to do it. I can
think of one issue, but it was already present in the script in the
first place: it's perfectly possible to have package managers from other
distributions on your system, for managing chroot-like things. Maybe
better OS detection or something more OS-independent would be better in
the longterm.